+++ B2COP-2D/src/Pfront.c
@@ -9,7 +9,7 @@ int cmp(const void* a, const void* b)
 }
 
 
-void Pfront_init(Pfront_t* bp, int size)
+void Pfront_init(Pfront_t* bp, size_t size)
 {
     bp->heapSize = 0;
     bp->keys = malloc(size * sizeof(my_m1));
@@ -17,7 +17,7 @@ void Pfront_init(Pfront_t* bp, int size)
     bp->maxSize = size;
 }
 
-void Pfront_zeroize(Pfront_t* bp, int size)
+void Pfront_zeroize(Pfront_t* bp, size_t size)
 {
     bp->heapSize = 0;
     memset(bp->keys, 0, size*sizeof(my_m1));
@@ -39,7 +39,7 @@ void Pfront_sort(Pfront_t* bp)
 }
 
 
-int Pfront_union(Pfront_t* pfront, Pfront_t* pfcand)
+size_t Pfront_union(Pfront_t* pfront, Pfront_t* pfcand)
 {
     if (pfront->heapSize + pfcand->heapSize > pfront->maxSize) {
         return pfront->maxSize;
@@ -51,7 +51,7 @@ int Pfront_union(Pfront_t* pfront, Pfront_t* pfcand)
 
 void Pfront_print(Pfront_t* pf, FILE* out)
 {
-    for (int i = 0 ; i < pf->heapSize ; i++) {
+    for (size_t i = 0 ; i < pf->heapSize ; i++) {
         fprintf(out, " %d ", pf->keys[i]);
     }
     printf("\n");
@@ -62,8 +62,8 @@ Pfront_t* Pfront_merge_sort(Pfront_t* pfront, Pfront_t* pfcand)
     Pfront_t* ret = malloc(sizeof(Pfront_t));
     Pfront_init(ret, pfront->maxSize);
 
-    int pfrontIndex = 0;
-    int pfcandIndex = 0;
+    size_t pfrontIndex = 0;
+    size_t pfcandIndex = 0;
 
     while (pfrontIndex < pfront->heapSize && pfcandIndex < pfcand->heapSize) {
         if (pfront->keys[pfrontIndex] < pfcand->keys[pfcandIndex]) {
@@ -100,9 +100,9 @@ void Pfront_free(Pfront_t* bh)
 }
 
 
-bool hasapath(Pfront_t** pfront, int dst)
+bool hasapath(Pfront_t** pfront, size_t dst)
 {
-    for (int i = 0 ; i <= SEG_MAX ; i++) {
+    for (size_t i = 0 ; i <= SEG_MAX ; i++) {
         if (pfront[i][dst].heapSize != 0) {
             return true;
         }
